backend/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
werkzeug
PyJWT==2.4.0